+++ circuits/src/field/decomposition/cpu_utils.rs
@@ -136,7 +136,7 @@ pub(super) fn process_limb_sizes(max_parallel_lookups: usize, limbs: &mut Vec<us
 ///             }
 ///         1 + OPT(bound - i*j)      
 ///      with SOL = [i; j] concatenated with SOL(bound - i*j)
-pub(super) fn compute_optimal_limb_sizes(
+pub(crate) fn compute_optimal_limb_sizes(
     solutions: &mut HashMap<i32, Vec<Vec<usize>>>,
     max_parallel_lookups: usize,
     max_bit_length: usize,