+++ zk_stdlib/tests/public_inputs.rs
@@ -117,7 +117,8 @@ fn pi_test(nb_public_inputs: u32, extra_pi: bool) {
     let mut srs = filecoin_srs(10);
 
     let relation = PIsCircuit { nb_public_inputs };
-    midnight_zk_stdlib::downsize_srs_for_relation(&mut srs, &relation);
+    let k = midnight_zk_stdlib::optimal_k(&relation);
+    srs.downsize(k);
     let vk = midnight_zk_stdlib::setup_vk(&srs, &relation);
     let pk = midnight_zk_stdlib::setup_pk(&relation, &vk);